feat: Rename experimental miekgdns2 resolver to miekgdns
Remove the old miekgdns resolver and replace it with the experimental
miekgdns2 resolver. The new resolver includes retry logic, connection
pooling, only uses TCP, and does not support search domains.
